+++ app/product/service/get_product.go
@@ -14,7 +14,7 @@ func GetProduct(ctx context.Context, req *rpcproduct.GetProductReq) (resp *rpcpr
 		return nil, errors.New("product id is required")
 	}
 	// get product by id
-	product, err := repository.GetProductById(ctx, uint(req.Id))
+	product, err := repository.GetProductById(ctx, req.Id)
 	if err != nil {
 		return
 	}
